backend/requirements.txt
chromadb==0.4.13
fastapi==0.108.0
langchain==0.1.0
langchain_community==0.0.11
langchain_openai==0.0.2
python-dotenv==1.0.0
python-multipart==0.0.6
pysqlite3-binary==0.5.2